major ui/ux optimizations and overhaul

+++ controller/commentsController.js
@@ -6,7 +6,7 @@ const { validationResult } = require("express-validator");
 exports.createComment = async (req, res, next) => {
   try {
     //create a new comment in the database
-    const { user, post, email, message } = req.body;
+    const { user, post, email, message, parentId } = req.body;
     //validate the request
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -19,13 +19,21 @@ exports.createComment = async (req, res, next) => {
             id: user,
           },
         },
-        email,
+        email: email, //email from the token
         message,
         post: {
           connect: {
             id: post,
           },
         },
+        //---------Testing a nested reply feature-----------
+        ...(parentId && {
+          parent: {
+            connect: {
+              id: parentId,
+            },
+          },
+        }),
       },
     });
     //respond with the created comment in the db
@@ -56,10 +64,15 @@ exports.getComments = async (req, res) => {
       include: {
         user: true,
         post: true,
+        replies: { // Include replies
+          include: {
+            user: true, // Include user details for replies
+          },
+        },
       },
       orderBy: {
         createdAt: "asc",
-        },
+      },
     });
 
     res.status(200).json({
@@ -128,6 +141,7 @@ exports.updateComment = async (req, res, next) => {
 exports.deleteComment = async (req, res, next) => {
   try {
     const commentId = req.params.id;
+    const userId = req.user.id;
     const comments = await prisma.comment.findUnique({
       where: {
         id: commentId,
@@ -135,6 +149,12 @@ exports.deleteComment = async (req, res, next) => {
     });
     if (!comments) {
       return res.status(404).json({ error: "the comment could not be found" });
+    }
+    // Check if the user is the owner of the comment
+    if (comments.userId !== userId) {
+      return res
+        .status(403)
+        .json({ error: "You are not authorized to delete this comment" });
     }
     await prisma.comment.delete({
       where: {
